# Conflicts:
#	backend/spring-routie/src/main/java/routie/routie/repository/RoutiePlaceRepository.java

+++ backend/spring-routie/src/main/java/routie/place/service/PlaceService.java
@@ -14,6 +14,7 @@ import routie.exception.BusinessException;
 import routie.exception.ErrorCode;
 import routie.place.repository.PlaceClosedDayOfWeekRepository;
 import routie.place.repository.PlaceRepository;
+import routie.routie.repository.RoutiePlaceRepository;
 import routie.routiespace.domain.RoutieSpace;
 import routie.routiespace.repository.RoutieSpaceRepository;
 
@@ -25,6 +26,7 @@ public class PlaceService {
     private final PlaceRepository placeRepository;
     private final RoutieSpaceRepository routieSpaceRepository;
     private final PlaceClosedDayOfWeekRepository placeClosedDayOfWeekRepository;
+    private final RoutiePlaceRepository routiePlaceRepository;
 
     public PlaceReadResponse getPlace(final String routieSpaceIdentifier, final long placeId) {
         final RoutieSpace routieSpace = getRoutieSpaceByIdentifier(routieSpaceIdentifier);
@@ -88,6 +90,9 @@ public class PlaceService {
         RoutieSpace routieSpace = getRoutieSpaceByIdentifier(routieSpaceIdentifier);
         if (!placeRepository.existsByIdAndRoutieSpace(placeId, routieSpace)) {
             throw new BusinessException(ErrorCode.PLACE_NOT_FOUND);
+        }
+        if (routiePlaceRepository.existsRoutiePlaceByPlaceId(placeId)) {
+            throw new BusinessException(ErrorCode.ROUTIE_PLACE_EXIST);
         }
         placeRepository.deleteById(placeId);
     }
